+++ BusinessEntity/SaleStatistic.cs
@@ -22,7 +22,7 @@ namespace BusinessEntity
 
         public string UnitName { get; set; }
 
-        public int Quanlity { get; set; }
+        public double Quanlity { get; set; }
 
         public double UnitPrice { get; set; }
